# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type BookFormat = typeof BookFormat[keyof typeof BookFormat];


export const BookFormat = {
  PDF: 'PDF',
  EPUB: 'EPUB',
} as const;

export interface Book {
  id: string;
  slug: string;
  title: string;
  author: string;
  price: number;
  currency: string;
  category: string;
  description: string;
  format: BookFormat;
  /** @nullable */
  coverUrl: string | null;
  fileName: string;
  featured: boolean;
  publishedAt: string;
  createdAt: string;
}

export type BookInputFormat = typeof BookInputFormat[keyof typeof BookInputFormat];


export const BookInputFormat = {
  PDF: 'PDF',
  EPUB: 'EPUB',
} as const;

export interface BookInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  slug: string;
  /** @minLength 1 */
  author: string;
  /** @minimum 0 */
  price: number;
  /**
     * @minLength 3
     * @maxLength 3
     */
  currency: string;
  /** @minLength 1 */
  category: string;
  description: string;
  format: BookInputFormat;
  /** @nullable */
  coverObjectPath: string | null;
  /** @minLength 1 */
  fileObjectPath: string;
  /** @minLength 1 */
  fileName: string;
  featured: boolean;
  publishedAt: string;
}

export type BookUpdateFormat = typeof BookUpdateFormat[keyof typeof BookUpdateFormat];


export const BookUpdateFormat = {
  PDF: 'PDF',
  EPUB: 'EPUB',
} as const;

export interface BookUpdate {
  /** @minLength 1 */
  title?: string;
  /** @minLength 1 */
  slug?: string;
  /** @minLength 1 */
  author?: string;
  /** @minimum 0 */
  price?: number;
  /**
     * @minLength 3
     * @maxLength 3
     */
  currency?: string;
  /** @minLength 1 */
  category?: string;
  description?: string;
  format?: BookUpdateFormat;
  /** @nullable */
  coverObjectPath?: string | null;
  /** @minLength 1 */
  fileObjectPath?: string;
  /** @minLength 1 */
  fileName?: string;
  featured?: boolean;
  publishedAt?: string;
}

export interface Category {
  name: string;
  count: number;
}

export interface StorefrontSummary {
  featured: Book[];
  newArrivals: Book[];
  categories: Category[];
}

export interface OrderInput {
  /** @minLength 3 */
  email: string;
  /**
     * @minLength 2
     * @maxLength 2
     */
  country: string;
  /**
     * @minLength 3
     * @maxLength 3
     */
  currency: string;
  /** @minItems 1 */
  bookIds: string[];
}

export interface CheckoutSession {
  orderId: string;
  reference: string;
  authorizationUrl: string;
  accessCode: string;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  paid: 'paid',
  failed: 'failed',
} as const;

export interface OrderItem {
  bookId: string;
  title: string;
  author: string;
  price: number;
  format: string;
  /** @nullable */
  downloadUrl: string | null;
  downloaded: boolean;
}

export interface Order {
  id: string;
  reference: string;
  email: string;
  country: string;
  currency: string;
  subtotal: number;
  status: OrderStatus;
  paymentStatus: string;
  deliveryEmailSent: boolean;
  downloaded: boolean;
  createdAt: string;
  /** @nullable */
  paidAt: string | null;
  items: OrderItem[];
}

export interface Download {
  bookId: string;
  title: string;
  format: string;
  url: string;
}

export type PaystackWebhookData = {
  reference: string;
  status?: string;
};

export interface PaystackWebhook {
  event: string;
  data: PaystackWebhookData;
}

export interface WebhookResponse {
  received: boolean;
}

export interface AdminLoginInput {
  /** @minLength 3 */
  email: string;
  /** @minLength 1 */
  password: string;
}

export interface AdminSession {
  authenticated: boolean;
  /** @nullable */
  email: string | null;
}

export interface AdminDashboard {
  totalOrders: number;
  pendingOrders: number;
  totalBooks: number;
  recentOrders: Order[];
}

export interface UploadInput {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadResponse {
  uploadURL: string;
  objectPath: string;
}

/**
 * Invalid request
 */
export type BadRequestResponse = ErrorResponse;

/**
 * Admin authentication required
 */
export type UnauthorizedResponse = ErrorResponse;

/**
 * Resource not found
 */
export type NotFoundResponse = ErrorResponse;

export type CategoryQueryParameter = string;

export type FormatQueryParameter = typeof FormatQueryParameter[keyof typeof FormatQueryParameter];


export const FormatQueryParameter = {
  PDF: 'PDF',
  EPUB: 'EPUB',
} as const;

export type SearchQueryParameter = string;

export type MaxPriceQueryParameter = number;

export type StatusQueryParameter = typeof StatusQueryParameter[keyof typeof StatusQueryParameter];


export const StatusQueryParameter = {
  pending: 'pending',
  paid: 'paid',
  failed: 'failed',
} as const;

export type ListBooksParams = {
category?: CategoryQueryParameter;
format?: FormatQueryParameter;
search?: SearchQueryParameter;
maxPrice?: MaxPriceQueryParameter;
};

export type ListAdminOrdersParams = {
status?: StatusQueryParameter;
};