# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /books:
    get:
      operationId: listBooks
      tags: [storefront]
      summary: List available ebooks
      parameters:
        - $ref: "#/components/parameters/CategoryQuery"
        - $ref: "#/components/parameters/FormatQuery"
        - $ref: "#/components/parameters/SearchQuery"
        - $ref: "#/components/parameters/MaxPriceQuery"
      responses:
        "200":
          description: Book catalogue
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Book"
  /books/{bookId}:
    get:
      operationId: getBook
      tags: [storefront]
      summary: Get a book
      parameters:
        - $ref: "#/components/parameters/BookIdPath"
      responses:
        "200":
          description: Book detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a private ebook or cover upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadInput"
      responses:
        "200":
          description: Presigned upload details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
components:
  parameters:
    BookIdPath:
      name: bookId
      in: path
      required: true
      schema:
        type: string
...
    CategoryQuery:
      name: category
      in: query
      required: false
      schema:
        type: string
    FormatQuery:
      name: format
      in: query
      required: false
      schema:
        type: string
        enum: [PDF, EPUB]
    SearchQuery:
      name: search
      in: query
      required: false
      schema:
        type: string
    MaxPriceQuery:
      name: maxPrice
      in: query
      required: false
      schema:
        type: number
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Admin authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    Book:
      type: object
      properties:
        id: { type: string }
        slug: { type: string }
        title: { type: string }
        author: { type: string }
        price: { type: number }
        currency: { type: string }
        category: { type: string }
        description: { type: string }
        format: { type: string, enum: [PDF, EPUB] }
        coverUrl: { type: ["string", "null"] }
        fileName: { type: string }
        featured: { type: boolean }
        publishedAt: { type: string }
        createdAt: { type: string }
      required: [id, slug, title, author, price, currency, category, description, format, coverUrl, fileName, featured, publishedAt, createdAt]
...
    UploadInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        size: { type: number, minimum: 1 }
        contentType: { type: string, minLength: 1 }
      required: [name, size, contentType]
    UploadResponse:
      type: object
      properties:
        uploadURL: { type: string }
        objectPath: { type: string }
      required: [uploadURL, objectPath]